Covered the basics
I spent about 40 minutes just to cover the bases here-  In real life this would take about 20 minutes to catch the big stuff and about 2-3 hours for the whole application. Again, I would want to do this upfront rather than after the fact- retro fitting and accommodation is not the best way to do accessibility- I understand the need to test me though.